Move "is this test code" related utilities to dedicated `IsUnitTestTrait`
The "is this test code" related utilities are only used by a small set of sniffs, so are better placed in a dedicated trait.

This moves the (`public`) `$custom_test_class_whitelist` and the `$custom_test_class_whitelist` properties as well as the `is_token_in_test_method()` and the `is_test_class` method to a new `WordPressCS\WordPress\Helpers\IsUnitTestTrait` and starts using that trait in the relevant sniffs.

The trait has been made stand-alone, i.e. it does not presume that a sniff implementing it extends the `WordPressCS\WordPress\Sniff` class.
This makes it easier to (re-)use the traits, even when a sniff extends an abstract sniff class from PHPCS itself or from PHPCSUtils.

Note: the `is_token_in_test_method()` method is currently unused in WPCS, but seems useful enough to keep in place.

Related to 1465