IsUnitTestTrait::is_token_in_test_method(): implement PHPCSUtils/improve handling nested functions
For the `is_token_in_test_method()` method, the functionality is changed in that the determination of whether something is a test method is more precise now.

Consider the below code sample:
```php
class MyUnitTest extends WP_UnitTestCase {
    public function MyTestHelper() {
        function Foo() {}
    }
}
```

The nested function `Foo` is not a method in the class, it is actually declared in the global namespace (or the applicable namespace if the file is namespaced), so code within that function should not be regarded as part of the test code.

This addressed that.

Note: this is a functional change which may need further discussion about what this function is supposed to check.

For example: for a token in a function nested within a test method, the method would previously return `true` and strictly speaking this is correct when the question is "is the token in a test method ?".

However, this answer is _incorrect_ if the question would be "is this a token which is unit test code ?" as the nested function would be in the global scope and not in the scope of the test class.