IsUnitTestTrait: add missing WP Core test case classes
Add two more test cases which were added to `/tests/phpunit/includes/`.

Note: WP Core contains more test cases, but those are in the `tests/phpunit/tests` directory, while the test cases in  `/tests/phpunit/includes/` are supposed to be the "public test API", which is why those are the only ones in this list.